requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
idna==2.8
mysql-connector-python==8.0.15
protobuf==3.7.1
requests==2.21.0
six==1.12.0
soupsieve==1.9
urllib3==1.24.1
